imbalance_research/requirements.txt
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
pyyaml>=6.0
tqdm>=4.65.0
tensorboard>=2.13.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
pynvml>=11.5.0
hydra-core>=1.3.0
omegaconf>=2.3.0
